naive conversion to nimble models
the functions do work (at least with the basic testing that I've done). However, I cannot for the life of me get these to work in BUGS code. Using nimble functions with vector bugs nodes is quirky. It might be that the vector has to be named? In any case, I'm nearing my wits end trying to figure out the million random uninformative errors NIMBLE throws at me. I also fixed a typo in the model functions. I could get it to work with individual inputs, but we don't want the model functions to be formatted like that.