simplified model eval functions
because nimble doesnt like doing the posterior predictives inside the function, it has to be done in the bugs code writing, not really a problem. Now, the posterior predictive argument in the model evaluation functions is useless, and is so removed. Honestly, I like this better since its more obvious what those functions do.